mock storage and api
