+++ crates/types/Cargo.toml
@@ -33,6 +33,9 @@ sha2          = { workspace = true }
 sha3          = { workspace = true, optional = true }
 thiserror     = { workspace = true }
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+grug-crypto = { path = "../crypto" }
+
 [dev-dependencies]
 proptest  = { workspace = true }
 test-case = { workspace = true }
