+++ crates/vm/rust/src/lib.rs
@@ -2,7 +2,6 @@ mod contract;
 mod error;
 #[rustfmt::skip]
 mod traits;
-mod api;
 mod vm;
 
-pub use {api::*, contract::*, error::*, traits::*, vm::*};
+pub use {contract::*, error::*, traits::*, vm::*};
